Move messages in right directory
