+++ coral/src/app/pages/topics/details/messages/index.tsx
@@ -1,4 +1,4 @@
-import { TopicMessages } from "src/app/features/topics/overview/messages/TopicMessages";
+import { TopicMessages } from "src/app/features/topics/details/messages/TopicMessages";
 
 function TopicMessagesPage() {
   return <TopicMessages />;
